Changes indicated above class names by 19.04.2023

+++ src/main/java/project/notizprogrammrepository/model/Types/segments/CalendarSegment.java
@@ -23,7 +23,7 @@ public class CalendarSegment extends Segment{
     }
 
     @Override
-    public Note getNote(int id) {
+    public Note getNote(double id) {
         return null;
     }
