+++ src/main/java/project/notizprogrammrepository/model/Types/segments/TodoSegment.java
@@ -23,7 +23,7 @@ public class TodoSegment extends Segment{
     }
 
     @Override
-    public Note getNote(int id) {
+    public Note getNote(double id) {
         return null;
     }
